refactor(ecs): keep quadrant id-encoding internal to @adobe/data
Drop quadrantOf/toLocalIndex from the public Entity namespace (leaving only
the isPersistent/isNonPersistent/isShared/isNonShared predicates). The load-time
quadrant bucketing that needed them moves into @adobe/data behind
serializedEntityLocationTables, so data-persistence hands over flat
(entity, archetype, row) triples and no longer touches the bit layout.
