+++ packages/data/src/ecs/entity/public.ts
@@ -1,3 +1,3 @@
 // © 2026 Adobe. MIT License. See /LICENSE for details.
 export { schema } from "./schema.js";
-export { isNonPersistent, isPersistent, isNonShared, isShared, quadrantOf, toLocalIndex } from "./persistence-sharing.js";
+export { isNonPersistent, isPersistent, isNonShared, isShared } from "./persistence-sharing.js";
